requirements.txt
flask
PyPDF2
google-generativeai
